예외처리

+++ src/main/java/com/sparta/boardwithcomment/controller/CommentController.java
@@ -3,12 +3,15 @@ package com.sparta.boardwithcomment.controller;
 import com.sparta.boardwithcomment.dto.CommentRequestDto;
 import com.sparta.boardwithcomment.dto.CommentResponseDto;
 import com.sparta.boardwithcomment.dto.CommentUpdateRequestDto;
+import com.sparta.boardwithcomment.entity.Posts;
 import com.sparta.boardwithcomment.security.UserDetailsImpl;
 import com.sparta.boardwithcomment.service.CommentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -19,17 +22,25 @@ public class CommentController {
     private final CommentService commentService;
 
     @PostMapping("/comment")
-    public CommentResponseDto save(@AuthenticationPrincipal UserDetailsImpl userDetails, @RequestBody CommentRequestDto requestDto){
-        return commentService.save(userDetails, requestDto);
+    public ResponseEntity<?> save(@AuthenticationPrincipal UserDetailsImpl userDetails, @RequestBody CommentRequestDto requestDto){
+        if (userDetails == null) return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("토큰이 유효하지 않습니다.");
+        return ResponseEntity.ok(commentService.save(userDetails, requestDto));
+
     }
 
     @PutMapping("/comment/{id}")
-    public CommentResponseDto update(@PathVariable Long id, @AuthenticationPrincipal UserDetailsImpl userDetails, @RequestBody CommentUpdateRequestDto requestDto){
-        return commentService.update(id, userDetails, requestDto);
+    public ResponseEntity<?> update(@PathVariable Long id, @AuthenticationPrincipal UserDetailsImpl userDetails, @RequestBody CommentUpdateRequestDto requestDto){
+        if (userDetails == null) return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("토큰이 유효하지 않습니다.");
+        try {
+            return ResponseEntity.ok(commentService.update(id, userDetails, requestDto));
+        } catch (Exception e){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
+        }
     }
 
     @DeleteMapping("/comment/{id}")
     public ResponseEntity<String> delete(@PathVariable Long id, @AuthenticationPrincipal UserDetailsImpl userDetails){
+        if (userDetails == null) return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("토큰이 유효하지 않습니다.");
         try{
             commentService.delete(id, userDetails);
             return ResponseEntity.status(HttpStatus.OK).body("댓글 삭제 성공");
@@ -37,5 +48,6 @@ public class CommentController {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
         }
     }
+
 
 }
